+++ src/Converters/Models/MatrixBlockType.php
@@ -4,6 +4,8 @@ namespace NerdsAndCompany\Schematic\Converters\Models;
 
 use Craft;
 use craft\base\Model;
+use craft\db\Query;
+use craft\models\MatrixBlockType as MatrixBlockTypeModel;
 
 /**
  * Schematic Matrix Block Types Converter.
@@ -38,13 +40,28 @@ class MatrixBlockType extends Base
      */
     public function saveRecord(Model $record, array $definition): bool
     {
-        $context = 'matrixBlockType:'.$record->id;
-        $existingFields = Craft::$app->fields->getAllFields($context);
+        // Get existing fields by block type handle
+        $existingBlockType = $this->getBlockTypeByHandle($definition['attributes']['handle']);
+        $existingFields = $existingBlockType ? $existingBlockType->getFields() : [];
+
+        // Set the content table for this matrix block
+        $originalContentTable = Craft::$app->content->contentTable;
+        $matrixField = Craft::$app->fields->getFieldById($record->fieldId);
+        $contentTable = Craft::$app->matrix->getContentTableName($matrixField);
+        Craft::$app->content->contentTable = $contentTable;
+
+        // Get the matrix block fields from the definition
         $modelMapper = Craft::$app->controller->module->modelMapper;
-        $fields = $modelMapper->import($definition['fields'], $existingFields, ['context' => $context], false);
+        $fields = $modelMapper->import($definition['fields'], $existingFields, [], false);
         $record->setFields($fields);
 
-        return Craft::$app->matrix->saveBlockType($record, false);
+        // Save the matrix block
+        $result = Craft::$app->matrix->saveBlockType($record, false);
+
+        // Restore the content table to what it was before
+        Craft::$app->content->contentTable = $originalContentTable;
+
+        return $result;
     }
 
     /**
@@ -53,5 +70,30 @@ class MatrixBlockType extends Base
     public function deleteRecord(Model $record): bool
     {
         return Craft::$app->matrix->deleteBlockType($record);
+    }
+
+    /**
+     * Get a Matrix Block Type Model by its handle.
+     *
+     * @param string $handle
+     *
+     * @return MatrixBlockTypeModel
+     */
+    private function getBlockTypeByHandle(string $handle): MatrixBlockTypeModel
+    {
+        $result = (new Query())
+            ->select([
+                'id',
+                'fieldId',
+                'fieldLayoutId',
+                'name',
+                'handle',
+                'sortOrder',
+            ])
+            ->from(['{{%matrixblocktypes}}'])
+            ->where(['handle' => $handle])
+            ->one();
+
+        return new MatrixBlockTypeModel($result);
     }
 }